SCRD-2082 Fix write_conf to avoid clash with ansible 2.x
In ansible 2.x "group" is a reserved action, so the code errors out
with "conflict between include and group action". Rename the parameters
to conf_group avoids this. As we don't really want to have configuration
files writeable for the service itself due to security concerns, we can
remove the owner parameter alltogether, significantly reducing
duplication.

Change-Id: I0b181965fed21536dc0a5f2136dd1b274f34e44e