Restart Nova compute after reboot on SLES failed (bsc#1086688)
The nova-compute service errored immediately after it started
after compute node reboots. The reason is that libvirtd failed to
start when nova-compute is started. The failure is caused
by the nove-compute.service file defined that it starts
after "libvirtd.target", which does not exist in SLES.
Also, libvirtd.target does not exist on RHEL.
The actual systemd definition file on SLES and RHEL
is "libvirtd.service".

Change-Id: I5796ed5ad971507fed8116ebe56f4363078af3fb