Create virsh secret for SES/ceph (bsc#1092579)
This patch migrates the code in the ses extension for setting up
the virsh secret key for communication with rbd.  The SES extension
runs prior to nova deploy, so virsh isn't installed yet.  So we
delegate the creation of the secret to the nova playbook after deploy
but as part of configure to ensure virsh exists.

Change-Id: I67a46cde1d49014df237f43f0a7479a7f56964eb