"qemu" user needs access to HugePage directory (bsc#1092431)
In ardana the qemu binary is run under the 'qemu' user
not root.
When using hugepages, the same user that runs the qemu
binary needs permissions to acces the directory where
hugepages are mounted.

It is best to mount hugepages in a more temporary and less
restrictive location.

We also have to give root access to the directory.

Change-Id: I1233c3dc146a7c5b8bc1b6bcfcb56fb6f0547eb0