Bypass setting hugepage directory file permissions (bsc#1092431)
There seems to be a defect/problem in ansible which
prevents it from using the mode option in the file module.
A quick search on "mode must be in octal or symbolic form"
turn up examples like:

https://github.com/ansible/ansible/issues/31952

The "mode" option is not required at
thispoint in the installation (nova deployment),
because the directory already has the correct mode
from os-config section of the installation. Thus
setting the hugepage directories file permissions
can be bypassed.

Change-Id: Ie19c5b655ad61ccbfba28ff8f0ad7c462e6b573b