Add ability to not set images_type (bsc#1113979)
This patch works with the SES integration to check to see if the
ses_nova_set_images_type boolean is set.  If the boolean is set to
False, then the hypervisor.conf doesn't set the images_type
in the [libvirt] section to rbd.  This allows the deployer to
maintain their own defined backend for images.  This still allows
nova to work with cinder ceph volumes.

refernces this:
https://gerrit.suse.provo.cloud/5263

Change-Id: I4b3195e4dcb0b3c3bd7a27094000ae74a888141d