enable all weigher classes by default (bsc#1123561)
We only enabled RAMWeigher by default which broke Magnum as it is
using the soft-anti-affinity policy for the server groups by default.
Also, only supporting RAMWeigher seem very inflexible. And from the
git change logs where wasn't any explanation why we placed that
restriction to begin with.

So in the absence of any good reasons to restrict to RAMWeigher,
and to unbreak Magnum, we should just remove that restriction and
use the Nove default, which is to load all weigher classes.

Change-Id: I60552523c8999652bb3c29e2a231836a9c35dda9
(cherry picked from commit ab67b7794f302a9120b280dc97a15ffc9ef3290a)