SCRD-7875 Remove explicit libvirt-python install (SCRD-7875)
As part of getting Ardana bootstrapped on SLE12 SP4 we had to add
the libvirt-python package to the nova compute nodes. We have since
addressed issues in the venv-openstack-nova packaging which should
resolve this problem in a better way and we shouldn't need to do
an explicit install of libvirt-python.

Change-Id: Ia231af620c3d7c053f27400bf39aa2a5822e5990