Disable nova-consoleauth monasca process check (bsc#1174242)
This patch removes the nova-consoleauth monasca process check
from nodes where it is not active. This is needed when you move the
nova-consoleauth (it is a singleton) from one controller to another.
Currently, the process check is not disabled and generates a false
alarm about the process not running. The "--remove" option for
monasca-setup is used to update the monasca-agent process check.

Change-Id: I6beb69f111619ecfbd2d2aca54ed86f293bfe1c0